refactor: move msword/ directory into resources/ directory
