+++ resources/msword/read.go
