+++ resources/msword/write.go
