+++ server/docx/docx.go
@@ -4,7 +4,7 @@ Package outputdata/docx handles writing Word Document (.DOCX) output.
 package docx
 
 import (
-	"github.com/Triangleman7/Interns_Summer_2022/msword"
+	"github.com/Triangleman7/Interns_Summer_2022/resources/msword"
 )
 
 // ReadTemplate reads the template Word Document located at path and returns the a Word Document
